move config generation back to implementation crate
